requirements.txt
certifi==2020.6.20
chardet==3.0.4
coverage==5.2.1
cprint==1.1
idna==2.10
Pillow==7.2.0
requests==2.24.0
six==1.15.0
termcolor==1.1.0
urllib3==1.25.10
vk-api==11.8.0
